refactor: extract participants upload/IO helpers into conversion_participants_io.py
Move 9 upload/IO helper functions and 2 module constants from the main blueprint to a focused IO module: _save_participants_upload_to_temp, _normalize_separator_option, _expected_delimiter_for_suffix, _read_participants_input_table, _get_excel_sheet_metadata, _resolve_participants_sheet_arg, _classify_time_style, _detect_mixed_time_style_columns, _format_mixed_time_style_message, plus _SUPPORTED_PARTICIPANTS_UPLOAD_SUFFIXES and _SUPPORTED_PARTICIPANTS_UPLOAD_MESSAGE. Update imports in the blueprint to use the new module.
